# server/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Sonarly API
  version: 2.0.0
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  description: >
    Native REST API of the Sonarly Go server. Documented routes
    are the native /api surface plus the /health, /healthz and /ready
    liveness probes;
    the OpenSubsonic adapter under /rest is a separate (Subsonic-compatible)
    contract and is intentionally not part of this spec.

    Error contract: every non-2xx response body is `{"error": "..."}` with a
    generic message for 5xx (driver errors never reach the client).
servers:
  - url: /
# Default auth: the signed sessionId cookie OR the X-API-Key header.
# Public operations override with `security: []`; /api/events overrides with
# session-cookie-only (EventSource cannot set headers).
security:
  - sessionId: []
  - apiKey: []
...
paths:
...
  /api/cover-art/{id}:
    get:
      operationId: getCoverArt
      tags: [catalog]
      summary: Cover art blob
      parameters:
        - $ref: '#/components/parameters/CoverArtId'
      responses:
        '200':
          description: The image bytes, private-cacheable for a day.
          headers:
            Cache-Control:
              schema: { type: string, example: 'private, max-age=86400' }
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List playlists visible to the caller
      description: >
        `shareToken` is populated only for the playlist's owner: other
        viewers authorize with it but never see it.
      responses:
        '200':
          description: Visible playlists (empty array when none).
          content:
            application/json:
              schema:
                type: object
                required: [playlists]
                properties:
                  playlists:
                    type: array
                    items: { $ref: '#/components/schemas/PlaylistListItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create a playlist (static, or smart when `rules` is present)
      description: >
        A body with `rules` creates a smart playlist; the old explicit
        `isSmart` flag is folded into rules presence. `visibility` defaults
        per service rules when omitted; `resolveMode` defaults to `tracks`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                visibility: { $ref: '#/components/schemas/PlaylistVisibility' }
                songIds:
                  type: array
                  items: { type: string }
                rules: { $ref: '#/components/schemas/SmartPlaylistRules' }
                resolveMode: { $ref: '#/components/schemas/ResolveMode' }
      responses:
        '201':
          description: Created playlist.
          content:
            application/json:
              schema:
                type: object
                required: [playlist]
                properties:
                  playlist: { $ref: '#/components/schemas/PlaylistDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Playlist detail with resolved entries
      description: >
        `shares` and `shareToken` are owner-only. The `shareToken` query
        parameter authorizes non-owner access to link-shared playlists.
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
        - { name: shareToken, in: query, schema: { type: string }, description: 'Link-share token (for viewers, not the owner).' }
      responses:
        '200':
          description: The playlist detail.
          content:
            application/json:
              schema:
                type: object
                required: [playlist]
                properties:
                  playlist: { $ref: '#/components/schemas/PlaylistDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updatePlaylist
      tags: [playlists]
      summary: Update a playlist (every field optional)
      description: >
        `rules` present (including explicit null) re-parses the smart rules;
        `songIds` present replaces the membership.
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                visibility: { $ref: '#/components/schemas/PlaylistVisibility' }
                songIds:
                  type: array
                  items: { type: string }
                isSmart: { type: boolean }
                rules:
                  anyOf:
                    - { $ref: '#/components/schemas/SmartPlaylistRules' }
                    - { type: 'null' }
                resolveMode: { $ref: '#/components/schemas/ResolveMode' }
      responses:
        '200':
          description: Updated playlist.
          content:
            application/json:
              schema:
                type: object
                required: [playlist]
                properties:
                  playlist: { $ref: '#/components/schemas/PlaylistDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - $ref: '#/components/parameters/PlaylistId'
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionId:
      type: apiKey
      in: cookie
      name: sessionId
      description: Signed session cookie (HttpOnly, SameSite=Strict), set by /api/login and /api/setup. The only credential /api/events accepts.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alternative to the session cookie on ordinary /api routes (not /api/events). SHA-256 digest looked up in api_keys.
  parameters:
...
    CoverArtId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Cover art id.
    PlaylistId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Playlist id (UUID).
...
  responses:
    BadRequest:
      description: Malformed input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: No or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not allowed (admin-gated route, or playlist access denied).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (or out of the caller's scope).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, enum: [true] }
    # --- enums ---------------------------------------------------------------
...
    PlaylistVisibility:
      type: string
      enum: [private, shared, public, link]
    ResolveMode:
      type: string
      enum: [tracks, query]
      description: >
        `tracks` resolves a smart playlist's user-scoped rule fields against
        the owner's data so every viewer receives the same curated list;
        `query` re-resolves live against each viewer's own data.
...
    NameEntry:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    PlaylistListItem:
      type: object
      description: >
        One entry of GET /api/playlists. `shareToken` is populated only for
        the playlist's owner (other viewers never see it).
      required: [id, name, ownerId, ownerUsername, visibility, isSmart, resolveMode, songCount, starred, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        ownerId: { type: string }
        ownerUsername: { type: string }
        visibility: { $ref: '#/components/schemas/PlaylistVisibility' }
        shareToken: { type: string }
        isSmart: { type: boolean }
        resolveMode: { $ref: '#/components/schemas/ResolveMode' }
        songCount: { type: integer }
        starred: { type: boolean }
        rating: { type: number }
        createdAt: { type: string }
        updatedAt: { type: string }
    PlaylistEntry:
      type: object
      description: >
        One song inside a playlist detail. `track`, `discNumber` and
        `duration` are always present but null when unknown; `album` and
        `artist` are the display names (empty string when unset), matching
        the old fetchPlaylistSongs shape.
      required: [id, title, album, artist, explicit, track, discNumber, duration, type, isDir, created]
      properties:
        id: { type: string }
        title: { type: string }
        album: { type: string }
        albumId: { type: string }
        artist: { type: string }
        artistId: { type: string }
        artists:
          type: array
          items: { type: string }
        artistEntries:
          type: array
          items: { $ref: '#/components/schemas/NameEntry' }
        track: { type: [integer, 'null'] }
        discNumber: { type: [integer, 'null'] }
        genre: { type: string }
        year: { type: integer }
        explicit: { type: boolean }
        duration: { type: [integer, 'null'], description: Seconds. }
        coverArt: { type: string }
        albumCoverArt: { type: string }
        type: { type: string }
        isDir: { type: boolean }
        created: { type: string }
    ShareEntry:
      type: object
      required: [userId, username, canEdit]
      properties:
        userId: { type: string }
        username: { type: string }
        canEdit: { type: boolean }
    PlaylistDetail:
      type: object
      description: '`shares` and `shareToken` are owner-only.'
      required: [id, name, ownerId, ownerUsername, visibility, isSmart, resolveMode, songCount, entries, starred, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        ownerId: { type: string }
        ownerUsername: { type: string }
        visibility: { $ref: '#/components/schemas/PlaylistVisibility' }
        shareToken: { type: string }
        isSmart: { type: boolean }
        resolveMode: { $ref: '#/components/schemas/ResolveMode' }
        songCount: { type: integer }
        entries:
          type: array
          items: { $ref: '#/components/schemas/PlaylistEntry' }
        shares:
          type: array
          items: { $ref: '#/components/schemas/ShareEntry' }
        starred: { type: boolean }
        rating: { type: number }
        createdAt: { type: string }
        updatedAt: { type: string }
    SmartPlaylistRules:
      type: object
      description: >
        Smart playlist rules (single-level groups, no nesting).
        A song matches when every `all` rule matches AND at least one `any`
        rule matches; a missing group side is not constrained.
        Field/operator whitelists are enforced by the rules compiler (see
        docs/smart-playlists.md), not by this schema.
      properties:
        rules:
          type: object
          properties:
            all:
              type: array
              items: { $ref: '#/components/schemas/SmartPlaylistRule' }
            any:
              type: array
              items: { $ref: '#/components/schemas/SmartPlaylistRule' }
        sort:
          type: array
          items: { $ref: '#/components/schemas/SmartPlaylistSort' }
        limit: { type: integer }
        limitPercent: { type: integer }
    SmartPlaylistRule:
      type: object
      required: [field, operator]
      properties:
        field: { type: string }
        operator: { type: string }
        value:
          description: 'String, number, boolean, or a two-element number array (inTheRange).'
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: array
              items: { type: number }
    SmartPlaylistSort:
      type: object
      description: 'Either `{field, direction}` or `{random: true}` — never both.'
      properties:
        field: { type: string }
        direction: { type: string, enum: [asc, desc] }
        random: { type: boolean }
    # --- playback ------------------------------------------------------------
...